+++ NEWS.md
@@ -1,4 +1,5 @@
 # spant 2.2.0
+* peak_info function now works with masked datasets.
 
 # spant 2.1.0
 * y-axis label can now be changed with a new argument (yaxis_lab) to