+++ NEWS.md
@@ -3,6 +3,8 @@
 * Bug fix for HSVD function not working with frequency domain data.
 * Bug fix for TWIX affine calculation when not using power of two matrix
 dimensions.
+* New function (crop_td_pts_pot) to round a dataset length down to the nearest
+power of 2.
 
 # spant 2.1.0
 * y-axis label can now be changed with a new argument (yaxis_lab) to